[Fixes #7860] moved set_thumbnail_from_bbox to resourcebase instead of dataset
