+++ geonode/base/api/views.py
@@ -16,6 +16,9 @@
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
 #########################################################################
+import ast
+from geonode.thumbs.thumbnails import create_thumbnail
+import json
 from django.apps import apps
 from django.conf import settings
 from django.db.models import Subquery
@@ -408,3 +411,31 @@ class ResourceBaseViewSet(DynamicModelViewSet):
         resource = self.get_object()
         resource.set_permissions(request.data)
         return Response(request.data)
+
+    @extend_schema(
+        methods=["post"], responses={200}, description="API endpoint allowing to set the thumbnail url for an existing dataset."
+    )
+    @action(
+        detail=False,
+        url_path="(?P<resource_id>\d+)/set_thumbnail_from_bbox",  # noqa
+        url_name="set-thumb-from-bbox",
+        methods=["post"],
+        permission_classes=[
+            IsAuthenticated,
+        ],
+    )
+    def set_thumbnail_from_bbox(self, request, resource_id):
+        try:
+            resource = ResourceBase.objects.get(id=ast.literal_eval(resource_id))
+            request_body = request.data if request.data else json.loads(request.body)
+            bbox = request_body["bbox"] + [request_body["srid"]]
+            zoom = request_body.get("zoom", None)
+
+            thumbnail_url = create_thumbnail(resource.get_real_instance(), bbox=bbox, background_zoom=zoom, overwrite=True)
+            return Response({"thumbnail_url": thumbnail_url}, status=200)
+        except ResourceBase.DoesNotExist:
+            logger.error(f"Resource selected with id {resource_id} does not exists")
+            return Response(data={"message": f"Resource selected with id {resource_id} does not exists"}, status=404, exception=True)
+        except Exception as e:
+            logger.error(e)
+            return Response(data={"message": e.args[0]}, status=500, exception=True)
