+++ geonode/layers/api/tests.py
@@ -16,10 +16,8 @@
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
 #########################################################################
-import sys
 import logging
 
-from unittest.mock import patch
 from urllib.parse import urljoin
 
 from django.urls import reverse
@@ -95,69 +93,3 @@ class DatasetsApiTests(APITestCase):
         self.assertEqual(response.data['dataset']['raw_constraints_other'], "None")
         self.assertEqual(response.data['dataset']['raw_supplemental_information'], "No information provided í £682m")
         self.assertEqual(response.data['dataset']['raw_data_quality_statement'], "OK    1 2   a b")
-
-    def test_datasets_set_thumbnail_from_bbox_from_Anonymous_user_raise_permission_error(self):
-        """
-        Given a request with Anonymous user, should raise an authentication error.
-        """
-        dataset_id = sys.maxsize
-        url = reverse('datasets-set-thumb-from-bbox', args=[dataset_id])
-        # Anonymous
-        expected = {
-            "detail": "Authentication credentials were not provided."
-        }
-        response = self.client.post(url, format='json')
-        self.assertEqual(response.status_code, 403)
-        self.assertEqual(expected, response.json())
-
-    @patch("geonode.layers.api.views.create_thumbnail")
-    def test_datasets_set_thumbnail_from_bbox_from_logged_user_for_existing_dataset(self, mock_create_thumbnail):
-        """
-        Given a logged User and an existing dataset, should create the expected thumbnail url.
-        """
-        mock_create_thumbnail.return_value = "http://localhost:8000/mocked_url.jpg"
-        # Admin
-        self.client.login(username="admin", password="admin")
-        dataset_id = Dataset.objects.first().resourcebase_ptr_id
-        url = reverse('datasets-set-thumb-from-bbox', args=[dataset_id])
-        payload = {
-            "bbox": [
-                -9072629.904175375,
-                -9043966.018568434,
-                1491839.8773032012,
-                1507127.2829602365
-            ],
-            "srid": "EPSG:3857"
-        }
-        response = self.client.post(url, data=payload, format='json')
-
-        expected = {
-            "thumbnail_url": "http://localhost:8000/mocked_url.jpg"
-        }
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(expected, response.json())
-
-    def test_datasets_set_thumbnail_from_bbox_from_logged_user_for_not_existing_dataset(self):
-        """
-        Given a logged User and an not existing dataset, should raise a 404 error.
-        """
-        # Admin
-        self.client.login(username="admin", password="admin")
-        dataset_id = sys.maxsize
-        url = reverse('datasets-set-thumb-from-bbox', args=[dataset_id])
-        payload = {
-            "bbox": [
-                -9072629.904175375,
-                -9043966.018568434,
-                1491839.8773032012,
-                1507127.2829602365
-            ],
-            "srid": "EPSG:3857"
-        }
-        response = self.client.post(url, data=payload, format='json')
-
-        expected = {
-            "message": f"Dataset selected with id {dataset_id} does not exists"
-        }
-        self.assertEqual(response.status_code, 404)
-        self.assertEqual(expected, response.json())
